requirements_rag.txt
langchain
langchain-community
langchain-chroma
langchain-huggingface
langchain-text-splitters
sentence-transformers
pypdf
chromadb
google-generativeai
